requirements.txt
telebot
python-dotenv
googletrans
pypinyin
